--- public/faq.css
@@ -1,22 +1,0 @@
-.faq {
-	background:   white;
-	color:        black;
-	display:      inline-block;
-	position:     absolute;
-	left:         50%;
-    box-shadow:   rgba(0, 0, 0, 0.14) 0px 6px 8px 1px, rgba(0, 0, 0, 0.12) 0px 3px 8px 2px, rgba(0, 0, 0, 0.3) 0px 4px 4px -3px;
-    margin-left:  -200px;
-    width:         400px;
-    background:    black;
-    color:         white;
-    height:        30px;
-    line-height:   30px;
-    padding-left:  5px;
-    padding-right: 5px;
-}
-.hidden {
-	display: none;
-}
-#bottompage {
-	height: 105px;
-}
