--- public/form.css
@@ -1,26 +1,0 @@
-* { box-sizing: border-box; }
-
-form {
-	margin: 0 auto;
-	width: 100%;
-	max-width: 500px;
-}
-
-input {
-	width: 100%;
-	margin-top: 10px;
-	border-radius: 3px;
-	border: 1px solid gray;
-	font-family: Verdana;
-	font-size: 20px;
-	outline: 0;
-}
-
-input[type="submit"]:hover {
-	background: rgb(250, 250, 250);
-	cursor: pointer;
-}
-
-.spacing {
-	height: 10px;
-}
