+++ public/maze.js
@@ -140,13 +140,13 @@ var mazeIntro = function(p) {
 
     Cell.prototype.highlight = function() {
         p.noFill();
-        p.stroke("#eb42f4");
+        p.stroke(255, 255, 255);
         p.ellipse(this.xPos + this.cellSize / 2, this.yPos + this.cellSize / 2, this.cellSize / 2, this.cellSize / 2);
-        p.fill(0, 0, 0);
+        p.fill(255, 255, 255);
     }
 
     Cell.prototype.mark = function() {
-        p.fill("#eb42f4");
+        p.fill(255, 255, 255);
         p.ellipse(this.xPos + this.cellSize / 2, this.yPos + this.cellSize / 2, this.cellSize / 2, this.cellSize / 2);
     }
